Initial restructure.
